+++ crates/types/src/externals.rs
@@ -1,4 +1,4 @@
-#![allow(unused_imports, unused_macros)]
+#![allow(deprecated, unused_imports, unused_macros)]
 
 use crate::{SchemaType, Schematic, StringType};
 
@@ -16,4 +16,37 @@ macro_rules! impl_string_format {
 }
 
 #[cfg(feature = "regex")]
-impl_string_format!(regex::Regex, "regex");
+mod regex_feature {
+    use super::*;
+
+    impl_string_format!(regex::Regex, "regex");
+}
+
+#[cfg(feature = "chrono")]
+mod chrono_feature {
+    use super::*;
+
+    macro_rules! impl_with_tz {
+        ($type:path, $format:expr) => {
+            impl<Tz: chrono::TimeZone> Schematic for $type {
+                fn generate_schema() -> SchemaType {
+                    SchemaType::String(StringType {
+                        format: Some($format.into()),
+                        ..StringType::default()
+                    })
+                }
+            }
+        };
+    }
+
+    impl_with_tz!(chrono::Date<Tz>, "date");
+    impl_with_tz!(chrono::DateTime<Tz>, "date-time");
+    impl_string_format!(chrono::Duration, "duration");
+    impl_string_format!(chrono::Days, "duration");
+    impl_string_format!(chrono::Months, "duration");
+    impl_string_format!(chrono::IsoWeek, "date");
+    impl_string_format!(chrono::NaiveWeek, "date");
+    impl_string_format!(chrono::NaiveDate, "date");
+    impl_string_format!(chrono::NaiveDateTime, "date-time");
+    impl_string_format!(chrono::NaiveTime, "time");
+}